# Conflicts:
#	elza/elza-core/src/main/java/cz/tacr/elza/bulkaction/AsyncBulkActionWorker.java
#	elza/elza-core/src/main/java/cz/tacr/elza/controller/ArrangementController.java
#	elza/elza-core/src/main/java/cz/tacr/elza/controller/vo/ap/item/ApItemUnitdateVO.java
#	elza/elza-core/src/main/java/cz/tacr/elza/core/data/StaticDataService.java
#	elza/elza-core/src/main/java/cz/tacr/elza/dataexchange/output/writer/xml/XmlSectionOutputStream.java
#	elza/elza-core/src/main/java/cz/tacr/elza/drools/ValidationRules.java
#	elza/elza-core/src/main/java/cz/tacr/elza/packageimport/PackageService.java
#	elza/elza-core/src/main/java/cz/tacr/elza/print/OutputModel.java
#	elza/elza-core/src/main/java/cz/tacr/elza/print/item/convertors/OutputItemConvertor.java
#	elza/elza-core/src/main/java/cz/tacr/elza/service/AccessPointDataService.java
#	elza/elza-core/src/main/java/cz/tacr/elza/service/AccessPointService.java
#	elza/elza-core/src/main/java/cz/tacr/elza/service/DescriptionItemService.java
#	elza/elza-core/src/main/java/cz/tacr/elza/service/StructObjService.java
#	elza/elza-core/src/main/java/cz/tacr/elza/service/importnodes/ImportProcess.java
#	elza/elza-core/src/test/java/cz/tacr/elza/ws/DaoCoreServiceTest.java
#	elza/elza-web/src/main/resources/frontend/actions/arr/node.jsx
#	elza/elza-web/src/main/resources/frontend/actions/arr/subNodeForm.jsx
#	elza/elza-web/src/main/resources/frontend/actions/router.jsx
#	elza/elza-web/src/main/resources/frontend/components/arr/FundBulkModificationsForm.jsx
#	elza/elza-web/src/main/resources/frontend/components/arr/SubNodeForm.jsx
#	elza/elza-web/src/main/resources/frontend/components/arr/nodeForm/DescItemLink.jsx
#	elza/elza-web/src/main/resources/frontend/components/arr/nodeForm/DescItemStructureRef.jsx
#	elza/elza-web/src/main/resources/frontend/components/arr/structure/UpdateMultipleSub.jsx
#	elza/elza-web/src/main/resources/frontend/components/form/ImportForm.jsx
#	elza/elza-web/src/main/resources/frontend/components/page/Ribbon.jsx
#	elza/elza-web/src/main/resources/frontend/components/shared/utils/itemFilter.jsx
#	elza/elza-web/src/main/resources/frontend/pages/AppRouter.jsx
#	elza/elza-web/src/main/resources/frontend/pages/Layout.jsx
#	elza/elza-web/src/main/resources/frontend/pages/arr/ArrDaoPage.jsx
#	elza/elza-web/src/main/resources/frontend/pages/arr/ArrPage.jsx
#	elza/elza-web/src/main/resources/frontend/pages/arr/ArrPageKeymap.jsx
#	elza/elza-web/src/main/resources/frontend/pages/node/NodePage.jsx
#	elza/elza-web/src/main/resources/frontend/stores/app/arr/node.jsx
#	elza/elza-web/src/main/resources/frontend/stores/app/arr/subNodeForm.jsx
#	elza/elza-web/src/main/resources/frontend/stores/app/router.jsx
#	elza/elza-web/src/main/resources/templates/web.html
#	elza/package-cz-base/src/ap_type.xml
#	elza/package-cz-base/src/package.xml
#	elza/rules-cz-zp2015/src/package.xml
#	elza/rules-cz-zp2015/src/rul_item_spec.xml
#	elza/rules-cz-zp2015/src/rul_item_type.xml
#	elza/rules-cz-zp2015/src/rul_structure_definition.xml
#	elza/rules-cz-zp2015/src/rul_structure_type.xml
#	elza/rules-simple-dev/src/package.xml

+++ elza/elza-core/src/main/java/cz/tacr/elza/packageimport/PackageService.java
@@ -32,6 +32,7 @@ import javax.xml.bind.JAXBContext;
 import javax.xml.bind.Marshaller;
 
 import cz.tacr.elza.repository.*;
+import cz.tacr.elza.service.AsyncRequestService;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang.BooleanUtils;
 import org.apache.commons.lang.StringUtils;
@@ -441,6 +442,9 @@ public class PackageService {
     @Autowired
     private IndexWorkProcessor indexWorkProcessor;
 
+    @Autowired
+    private AsyncRequestService asyncRequestService;
+
     /**
      * Provede import balíčku.
      *
@@ -500,7 +504,7 @@ public class PackageService {
             }
         }
 
-        if (oldPackageDir != null) {
+        if (oldPackageDir != null && !oldPackageDir.equals(packageDir)) {
             FileSystemUtils.deleteRecursively(oldPackageDir);
         }
 
@@ -525,11 +529,16 @@ public class PackageService {
         // spustit indexovani
         indexWorkProcessor.resumeIndexing();
 
+        asyncRequestService.start();
+
         logger.info("All async threads started.");
     }
 
     public void stopAsyncTasks() {
         logger.debug("Stopping async threads...");
+
+        asyncRequestService.stop();
+
         // zastavit indexovani
         indexWorkProcessor.suspendIndexing();
